# docs/swagger.yaml
definitions:
...
  main.MovieShow:
    properties:
      hall_id:
        type: string
      id:
        type: string
      language:
        type: string
      movie_id:
        type: string
      start_time:
        description: формат "HH:MM:SS"
        type: string
    type: object
  main.Review:
    properties:
      id:
        type: string
      movie_id:
        type: string
      rating:
        type: number
      review_comment:
        type: string
      user_id:
        type: string
    type: object
  main.Seat:
    properties:
      hall_id:
        type: string
      id:
        type: string
      row_number:
        type: integer
      seat_number:
        type: integer
      seat_type_id:
        type: string
    type: object
...
  main.Ticket:
    properties:
      id:
        type: string
      movie_show_id:
        type: string
      price:
        type: number
      seat_id:
        type: string
      ticket_status_id:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a sample
  title: Your API Title
  version: "1.0"
paths:
...
  /movie-shows:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.MovieShow'
            type: array
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Получить все показы фильмов
      tags:
      - movie-shows
    post:
      consumes:
      - application/json
      parameters:
      - description: Показ фильма
        in: body
        name: movie_show
        required: true
        schema:
          $ref: '#/definitions/main.MovieShow'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.MovieShow'
        "400":
          description: Неверный запрос
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Создать показ фильма
      tags:
      - movie-shows
  /movie-shows/{id}:
    delete:
      parameters:
      - description: ID показа фильма
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Удалено
          schema:
            type: string
        "404":
          description: Показ не найден
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Удалить показ фильма
      tags:
      - movie-shows
    get:
      parameters:
      - description: ID показа фильма
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.MovieShow'
        "404":
          description: Показ фильма не найден
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Получить показ фильма по ID
      tags:
      - movie-shows
    put:
      consumes:
      - application/json
      parameters:
      - description: ID показа фильма
        in: path
        name: id
        required: true
        type: string
      - description: Обновлённые данные показа
        in: body
        name: movie_show
        required: true
        schema:
          $ref: '#/definitions/main.MovieShow'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.MovieShow'
        "400":
          description: Неверный JSON
          schema:
            type: string
        "404":
          description: Показ не найден
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Обновить показ фильма
      tags:
      - movie-shows
...
  /reviews:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Review'
            type: array
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Получить все отзывы
      tags:
      - reviews
    post:
      consumes:
      - application/json
      parameters:
      - description: Новый отзыв
        in: body
        name: review
        required: true
        schema:
          $ref: '#/definitions/main.Review'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.Review'
        "400":
          description: Неверный запрос
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Создать отзыв
      tags:
      - reviews
  /reviews/{id}:
    delete:
      parameters:
      - description: ID отзыва
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Удалено
          schema:
            type: string
        "404":
          description: Отзыв не найден
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Удалить отзыв
      tags:
      - reviews
    get:
      parameters:
      - description: ID отзыва
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Review'
        "404":
          description: Отзыв не найден
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Получить отзыв по ID
      tags:
      - reviews
    put:
      consumes:
      - application/json
      parameters:
      - description: ID отзыва
        in: path
        name: id
        required: true
        type: string
      - description: Обновленные данные отзыва
        in: body
        name: review
        required: true
        schema:
          $ref: '#/definitions/main.Review'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Review'
        "400":
          description: Неверный запрос
          schema:
            type: string
        "404":
          description: Отзыв не найден
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Обновить отзыв
      tags:
      - reviews
...
  /seats:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Seat'
            type: array
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Получить все места
      tags:
      - seats
    post:
      consumes:
      - application/json
      parameters:
      - description: Новое место
        in: body
        name: seat
        required: true
        schema:
          $ref: '#/definitions/main.Seat'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.Seat'
      summary: Создать место
      tags:
      - seats
  /seats/{id}:
    delete:
      parameters:
      - description: ID места
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Удалено
          schema:
            type: string
        "404":
          description: Место не найдено
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Удалить место
      tags:
      - seats
    get:
      parameters:
      - description: ID места
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Seat'
        "404":
          description: Место не найдено
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Получить место по ID
      tags:
      - seats
    put:
      consumes:
      - application/json
      parameters:
      - description: ID места
        in: path
        name: id
        required: true
        type: string
      - description: Обновлённые данные места
        in: body
        name: seat
        required: true
        schema:
          $ref: '#/definitions/main.Seat'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Seat'
        "400":
          description: Неверный JSON
          schema:
            type: string
        "404":
          description: Место не найдено
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Обновить место
      tags:
      - seats
  /seats/{movie_show_id}:
    get:
      parameters:
      - description: ID показа фильма
        in: path
        name: movie_show_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Ticket'
            type: array
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Получить все билеты для показа фильма по ID
      tags:
      - seats
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"